파일 입출력 및 도감
1. 파일 입출력으로 보유중인 아이템 바이너리로 저장함.
2. 불러온 아이템 정보를 바탕으로 도감 작성함.
3. Stan Asset 적용시킴.

+++ Assets/Script/Sale.cs
@@ -7,7 +7,7 @@ public class Sale : MonoBehaviour
         if (col.tag == "Material")
         {
             //Debug.Log("Trigger");
-            DataController.GetInstance().AddGold(col.GetComponent<ItemInfo>().sellPrice);
+            DataController.GetInstance().AddGold((ulong)col.GetComponent<ItemInfo>().sellPrice);
             DataController.GetInstance().SubItemCount();
 
             Destroy(col.gameObject);
